requirements.txt
Flask
twitter-mood
python-twitter
